PYIC-4120: Update test case with Lime feedback and split into two contracts.
