+++ lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/passportCri/CredentialTests.java
@@ -10,12 +10,8 @@ import au.com.dius.pact.core.model.annotations.Pact;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nimbusds.jose.JOSEException;
-import com.nimbusds.jose.JWSAlgorithm;
 import com.nimbusds.jose.JWSSigner;
-import com.nimbusds.jose.util.Base64URL;
 import com.nimbusds.oauth2.sdk.http.HTTPResponse;
-import com.nimbusds.oauth2.sdk.token.AccessTokenType;
 import com.nimbusds.oauth2.sdk.token.BearerAccessToken;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Disabled;
@@ -24,7 +20,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.di.ipv.core.library.config.ConfigurationVariable;
 import uk.gov.di.ipv.core.library.domain.ContraIndicatorConfig;
 import uk.gov.di.ipv.core.library.domain.ErrorResponse;
 import uk.gov.di.ipv.core.library.dto.CriCallbackRequest;
@@ -48,13 +43,10 @@ import java.time.Instant;
 import java.time.ZoneOffset;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Set;
 
 import static au.com.dius.pact.consumer.dsl.LambdaDsl.newJsonBody;
 import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.greaterThan;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
@@ -63,9 +55,9 @@ import static org.mockito.Mockito.when;
 @Disabled("PACT tests should not be run in build pipelines at this time")
 @ExtendWith(PactConsumerTestExt.class)
 @ExtendWith(MockitoExtension.class)
-@PactTestFor(providerName = "PassportCriProvider")
+@PactTestFor(providerName = "PassportVcProvider")
 @MockServerConfig(hostInterface = "localhost", port = "1234")
-class ContractTest {
+class CredentialTests {
     private static final String TEST_USER = "test-subject";
     private static final String IPV_CORE_CLIENT_ID = "ipv-core";
     private static final String PRIVATE_API_KEY = "dummyApiKey";
@@ -81,13 +73,6 @@ class ContractTest {
             {"kty":"RSA","e":"AQAB","n":"vyapkvJXLwpYRJjbkQD99V2gcPEUKrO3dwjcAA9TPkLucQEZvYZvb7-wfSHxlvJlJcdS20r5PKKmqdPeW3Y4ir3WsVVeiht2iOZUreUO5O3V3o7ImvEjPS_2_ZKMHCwUf51a6WGOaDjO87OX_bluV2dp01n-E3kiIl6RmWCVywjn13fX3jsX0LMCM_bt3HofJqiYhhNymEwh39oR_D7EE5sLUii2XvpTYPa6L_uPwdKa4vRl4h4owrWEJaJifMorGcvqhCK1JOHqgknN_3cb_ns9Px6ynQCeFXvBDJy4q71clkBq_EZs5227Y1S222wXIwUYN8w5YORQe3M-pCIh1Q"}
             """;
     private static final ObjectMapper objectMapper = new ObjectMapper();
-
-    private static final String CLIENT_ASSERTION_HEADER = "eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiJ9";
-    private static final String CLIENT_ASSERTION_BODY =
-            "eyJpc3MiOiJpcHYtY29yZSIsInN1YiI6Imlwdi1jb3JlIiwiYXVkIjoiZHVtbXlQYXNzcG9ydENvbXBvbmVudElkIiwiZXhwIjo0MDcwOTA5NzAwLCJqdGkiOiJTY25GNGRHWHRoWllYU181azg1T2JFb1NVMDRXLUgzcWFfcDZucHYyWlVZIn0";
-    // Signature generated using JWT.io
-    private static final String CLIENT_ASSERTION_SIGNATURE =
-            "8_p0JKL8REQI7GXwqMs2qdKyDNstjUDv0RUgZc37LmdtFvIib317AbSbC9tM_J_YXI6pmhksuLvZJ_3fFNOJtg";
 
     // We hardcode the VC headers and bodies like this so that it is easy to update them from JSON
     // sent by the CRI team
@@ -286,6 +271,12 @@ class ContractTest {
                             "ci": [
                                 "CI01"
                             ],
+                            "checkDetails": [
+                                {
+                                    "checkMethod": "data",
+                                    "dataCheck": "record_check"
+                                }
+                            ],
                             "failedCheckDetails": [
                                 {
                                     "checkMethod": "data",
@@ -294,12 +285,6 @@ class ContractTest {
                                 {
                                     "checkMethod": "data",
                                     "dataCheck": "scenario2_check"
-                                }
-                            ],
-                            "checkDetails": [
-                                {
-                                    "checkMethod": "data",
-                                    "dataCheck": "record_check"
                                 }
                             ],
                             "ciReasons": [
@@ -317,154 +302,13 @@ class ContractTest {
     // valid signature (using https://jwt.io works well) and record it here so the PACT file doesn't
     // change each time we run the tests.
     private static final String FAILED_VC_SCENARIO_2_SIGNATURE =
-            "bzGrZaM60pussX6G9Jbrlyi15PAJU0J7Z8YhXDjfgfE_7-H0JuZDBde_OGr_pR3jJ72-jTyr44QO0xv0c6LT4Q";
+            "Ybc2oCiLSWA_Fi1mCrl5W9PHsjqpOOVlW_G507657EApUQIApb1wl1hG-v8Sip8TUwKCuhDBU-W4q3NW56tW8g";
 
     @Mock private ConfigService mockConfigService;
     @Mock private JWSSigner mockSigner;
     @Mock private SecureTokenHelper mockSecureTokenHelper;
 
-    @Pact(provider = "PassportCriProvider", consumer = "IpvCoreBack")
-    public RequestResponsePact validRequestReturnsValidAccessToken(PactDslWithProvider builder) {
-        return builder.given("dummyAuthCode is a valid authorization code")
-                .given("dummyApiKey is a valid api key")
-                .given("dummyPassportComponentId is the passport CRI component ID")
-                .given(
-                        "Passport CRI uses CORE_BACK_SIGNING_PRIVATE_KEY_JWK to validate core signatures")
-                .uponReceiving("Valid auth code")
-                .path("/token")
-                .method("POST")
-                .body(
-                        "client_assertion_type=urn%3Aietf%3Aparams%3Aoauth%3Aclient-assertion-type%3Ajwt-bearer&code=dummyAuthCode&grant_type=authorization_code&redirect_uri=https%3A%2F%2Fidentity.staging.account.gov.uk%2Fcredential-issuer%2Fcallback%3Fid%3DukPassport&client_assertion="
-                        + CLIENT_ASSERTION_HEADER + "." + CLIENT_ASSERTION_BODY + "." + CLIENT_ASSERTION_SIGNATURE)
-                .headers(
-                        "x-api-key",
-                        PRIVATE_API_KEY,
-                        "Content-Type",
-                        "application/x-www-form-urlencoded; charset=UTF-8")
-                .willRespondWith()
-                .status(200)
-                .body(
-                        newJsonBody(
-                                        (body) -> {
-                                            body.stringType("access_token");
-                                            body.stringValue("token_type", "Bearer");
-                                            body.integerType("expires_in");
-                                        })
-                                .build())
-                .toPact();
-    }
-
-    @Test
-    @PactTestFor(pactMethod = "validRequestReturnsValidAccessToken")
-    void fetchAccessToken_whenCalledAgainstPassportCri_retrievesAValidAccessToken(
-            MockServer mockServer) throws URISyntaxException, JOSEException, CriApiException {
-        // Arrange
-        var credentialIssuerConfig = getMockCredentialIssuerConfig(mockServer);
-
-        when(mockConfigService.getSsmParameter(ConfigurationVariable.JWT_TTL_SECONDS))
-                .thenReturn("900");
-        when(mockConfigService.getOauthCriConfig(any())).thenReturn(credentialIssuerConfig);
-        when(mockConfigService.getCriPrivateApiKey(any())).thenReturn(PRIVATE_API_KEY);
-
-        // Signature generated by jwt.io by debugging the test and getting the client assertion JWT
-        // generated by the test as mocking out the AWSKMS class inside the real signer would be
-        // painful.
-        when(mockSigner.sign(any(), any()))
-                .thenReturn(
-                        new Base64URL(
-                                CLIENT_ASSERTION_SIGNATURE));
-        when(mockSigner.supportedJWSAlgorithms()).thenReturn(Set.of(JWSAlgorithm.ES256));
-        when(mockSecureTokenHelper.generate())
-                .thenReturn("ScnF4dGXthZYXS_5k85ObEoSU04W-H3qa_p6npv2ZUY");
-
-        // We need to generate a fixed request, so we set the secure token and expiry to constant
-        // values.
-        var underTest =
-                new CriApiService(
-                        mockConfigService, mockSigner, mockSecureTokenHelper, CURRENT_TIME);
-
-        // Act
-        BearerAccessToken accessToken =
-                underTest.fetchAccessToken(
-                        getCallbackRequest("dummyAuthCode", credentialIssuerConfig),
-                        getCriOAuthSessionItem());
-        // Assert
-        assertThat(accessToken.getType(), is(AccessTokenType.BEARER));
-        assertThat(accessToken.getValue(), notNullValue());
-        assertThat(accessToken.getLifetime(), greaterThan(0L));
-    }
-
-    @Pact(provider = "PassportCriProvider", consumer = "IpvCoreBack")
-    public RequestResponsePact invalidAuthCodeRequestReturns400(PactDslWithProvider builder) {
-        return builder.given("dummyInvalidAuthCode is an invalid authorization code")
-                .given("dummyApiKey is a valid api key")
-                .given("dummyPassportComponentId is the passport CRI component ID")
-                .given(
-                        "Passport CRI uses CORE_BACK_SIGNING_PRIVATE_KEY_JWK to validate core signatures")
-                .uponReceiving("Invalid authorization code")
-                .path("/token")
-                .method("POST")
-                .body(
-                        "client_assertion_type=urn%3Aietf%3Aparams%3Aoauth%3Aclient-assertion-type%3Ajwt-bearer&code=dummyInvalidAuthCode&grant_type=authorization_code&redirect_uri=https%3A%2F%2Fidentity.staging.account.gov.uk%2Fcredential-issuer%2Fcallback%3Fid%3DukPassport&client_assertion="
-                                + CLIENT_ASSERTION_HEADER + "." + CLIENT_ASSERTION_BODY + "." + CLIENT_ASSERTION_SIGNATURE)
-                .headers(
-                        "x-api-key",
-                        PRIVATE_API_KEY,
-                        "Content-Type",
-                        "application/x-www-form-urlencoded; charset=UTF-8")
-                .willRespondWith()
-                .status(400)
-                .toPact();
-    }
-
-    @Test
-    @PactTestFor(pactMethod = "invalidAuthCodeRequestReturns400")
-    void fetchAccessToken_whenCalledAgainstPassportCriWithInvalidAuthCode_throwsAnException(
-            MockServer mockServer) throws URISyntaxException, JOSEException {
-
-        // Arrange
-        var credentialIssuerConfig = getMockCredentialIssuerConfig(mockServer);
-
-        when(mockConfigService.getSsmParameter(ConfigurationVariable.JWT_TTL_SECONDS))
-                .thenReturn("900");
-        when(mockConfigService.getOauthCriConfig(any())).thenReturn(credentialIssuerConfig);
-        when(mockConfigService.getCriPrivateApiKey(any())).thenReturn(PRIVATE_API_KEY);
-
-        // Signature generated by jwt.io by debugging the test and getting the client assertion
-        // JWT
-        // generated by the test as mocking out the AWSKMS class inside the real signer would be
-        // painful.
-        when(mockSigner.sign(any(), any()))
-                .thenReturn(
-                        new Base64URL(
-                                CLIENT_ASSERTION_SIGNATURE));
-        when(mockSigner.supportedJWSAlgorithms()).thenReturn(Set.of(JWSAlgorithm.ES256));
-        when(mockSecureTokenHelper.generate())
-                .thenReturn("ScnF4dGXthZYXS_5k85ObEoSU04W-H3qa_p6npv2ZUY");
-
-        // We need to generate a fixed request, so we set the secure token and expiry to
-        // constant
-        // values.
-        var underTest =
-                new CriApiService(
-                        mockConfigService, mockSigner, mockSecureTokenHelper, CURRENT_TIME);
-
-        // Act
-        CriApiException exception =
-                assertThrows(
-                        CriApiException.class,
-                        () ->
-                                underTest.fetchAccessToken(
-                                        getCallbackRequest(
-                                                "dummyInvalidAuthCode", credentialIssuerConfig),
-                                        getCriOAuthSessionItem()));
-
-        // Assert
-        assertEquals("Invalid token request", exception.getErrorResponse().getMessage());
-        assertEquals(400, exception.getHttpStatusCode());
-    }
-
-    @Pact(provider = "PassportCriProvider", consumer = "IpvCoreBack")
+    @Pact(provider = "PassportVcProvider", consumer = "IpvCoreBack")
     public RequestResponsePact validRequestReturnsValidCredential(PactDslWithProvider builder)
             throws Exception {
         return builder.given("dummyApiKey is a valid api key")
@@ -477,7 +321,7 @@ class ContractTest {
                 .given("VC passport documentNumber is 824159121")
                 .given("VC passport expiryDate is 2030-01-01")
                 .uponReceiving("Valid credential request for VC")
-                .path("/credential")
+                .path("/issue/credential")
                 .method("POST")
                 .headers("x-api-key", PRIVATE_API_KEY, "Authorization", "Bearer dummyAccessToken")
                 .willRespondWith()
@@ -548,7 +392,7 @@ class ContractTest {
                         });
     }
 
-    @Pact(provider = "PassportCriProvider", consumer = "IpvCoreBack")
+    @Pact(provider = "PassportVcProvider", consumer = "IpvCoreBack")
     public RequestResponsePact validRequestReturnsFailedCredentialWithCi(
             PactDslWithProvider builder) throws Exception {
         return builder.given("dummyApiKey is a valid api key")
@@ -561,7 +405,7 @@ class ContractTest {
                 .given("VC passport documentNumber is 123456789")
                 .given("VC passport expiryDate is 2030-12-12")
                 .uponReceiving("Valid credential request for VC with CI")
-                .path("/credential")
+                .path("/issue/credential")
                 .method("POST")
                 .headers("x-api-key", PRIVATE_API_KEY, "Authorization", "Bearer dummyAccessToken")
                 .willRespondWith()
@@ -637,7 +481,7 @@ class ContractTest {
                         });
     }
 
-    @Pact(provider = "PassportCriProvider", consumer = "IpvCoreBack")
+    @Pact(provider = "PassportVcProvider", consumer = "IpvCoreBack")
     public RequestResponsePact validRequestReturnsFailedCredentialWithScenario2Ci(
             PactDslWithProvider builder) throws Exception {
         return builder.given("dummyApiKey is a valid api key")
@@ -651,7 +495,7 @@ class ContractTest {
                 .given("VC passport expiryDate is 2030-12-12")
                 .given("VC is a scenario 2 failure")
                 .uponReceiving("Valid credential request for VC with scenario 2 CI")
-                .path("/credential")
+                .path("/issue/credential")
                 .method("POST")
                 .headers("x-api-key", PRIVATE_API_KEY, "Authorization", "Bearer dummyAccessToken")
                 .willRespondWith()
@@ -727,14 +571,14 @@ class ContractTest {
                         });
     }
 
-    @Pact(provider = "PassportCriProvider", consumer = "IpvCoreBack")
+    @Pact(provider = "PassportVcProvider", consumer = "IpvCoreBack")
     public RequestResponsePact invalidAccessTokenReturns403(PactDslWithProvider builder) {
         return builder.given("dummyApiKey is a valid api key")
                 .given("dummyInvalidAccessToken is an invalid access token")
                 .given("test-subject is a valid subject")
                 .given("dummyPassportComponentId is a valid issuer")
                 .uponReceiving("Invalid credential request due to invalid access token")
-                .path("/credential")
+                .path("/issue/credential")
                 .method("POST")
                 .headers(
                         "x-api-key",
@@ -840,7 +684,7 @@ class ContractTest {
             throws URISyntaxException {
         return OauthCriConfig.builder()
                 .tokenUrl(new URI("http://localhost:" + mockServer.getPort() + "/token"))
-                .credentialUrl(new URI("http://localhost:" + mockServer.getPort() + "/credential"))
+                .credentialUrl(new URI("http://localhost:" + mockServer.getPort() + "/issue/credential"))
                 .authorizeUrl(new URI("http://localhost:" + mockServer.getPort() + "/authorize"))
                 .clientId(IPV_CORE_CLIENT_ID)
                 .signingKey(CRI_SIGNING_PRIVATE_KEY_JWK)
